requirements.txt
django==4.2.6
gunicorn==21.2.0
psycopg2-binary==2.9.9
ipython==8.16.1
ipdb==0.13.13
simplejson==3.19.2
structlog==23.2.0
requests==2.31.0
django-google-sso==3.3.0
celery==5.3.4
openai==0.28.1
gcsa==2.1.0
django-tinymce==3.6.1
BeautifulSoup4==4.12.2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
djangorestframework
django-celery-beat